+++ workshop/spring-data/r2dbc-coroutines/src/main/kotlin/io/bluetape4k/workshop/r2dbc/domain/PostRepository.kt
@@ -59,7 +59,7 @@ class PostRepository(
             .awaitFirstOrNull()
     }
 
-    suspend fun deleteAll(): Int {
+    suspend fun deleteAll(): Long {
         return operations.delete<Post>().allAndAwait()
     }
